requirements.txt
accelerate
aiofiles
altair
annotated-types
anyio
attrs
backports.tarfile
blobfile
build
certifi
charset-normalizer
click
clip @ git+https://github.com/openai/CLIP.git
colorama
contourpy
cycler
diffusers
dnspython
docutils
email_validator
exceptiongroup
fastapi
fastapi-cli
ffmpy
filelock
fire
fonttools
fsspec
ftfy
gradio==4.37.2
gradio_client
h11
httpcore
httptools
httpx
huggingface-hub
humanize
idna
imageio
importlib_metadata
importlib_resources
intel-openmp
ipywidgets
jaraco.classes
jaraco.context
jaraco.functools
Jinja2
jsonschema
jsonschema-specifications
keyring
kiwisolver
lazy_loader
lxml
markdown-it-py
MarkupSafe
matplotlib
mdurl
mkl
more-itertools
mpmath
networkx
nh3
numpy
opencv-python
orjson
packaging
pandas
pillow
pkginfo
psutil
pycryptodomex
pydantic
pydantic_core
pydub
Pygments
pyparsing
pyproject_hooks
python-dateutil==2.9.0.post0
python-dotenv
python-multipart
pytz
pywin32-ctypes
PyYAML
readme_renderer
referencing
regex
requests
requests-toolbelt
rfc3986
rich
rpds-py
ruff
safetensors
scikit-image
scipy
semantic-version
-e git+https://github.com/openai/shap-e@50131012ee11c9d2617f3886c10f000d3c7a3b43#egg=shap_e
shellingham
six
sniffio
starlette
sympy
tbb
termcolor
tifffile
tokenizers
tomli
tomlkit
toolz
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tqdm
transformers
twine
typer
typing_extensions
tzdata
ujson
urllib3
uvicorn
watchfiles
wcwidth
websockets
zipp
